Unify into ConferenceDisplay component
